Stack_final

+++ TestStack.cpp
@@ -7,7 +7,7 @@ int main()
     int elements_in_stack = 10;
     Stack new_stack;
 
-    for (size_t i = 0; i < elements_in_stack; i++)
+    for (int i = 0; i < elements_in_stack; i++)
     {
         new_stack.push(i);
     }
@@ -15,11 +15,10 @@ int main()
     std::cout << std::boolalpha;
     std::cout << "Is new stack empty? - " << new_stack.isEmpty() << std::endl;
     
-    for (size_t i = 0; i < elements_in_stack; i++)
+    for (int i = 0; i < elements_in_stack; i++)
     {
         new_stack.pop();
     }
-
     //new_stack.pop();
 
     std::cout << "Is new stack empty? - " << new_stack.isEmpty() << std::endl;
